renamed information processor to information processor abstract to disambiguate with package
