+++ clayrs/content_analyzer/field_content_production_techniques/embedding_technique/embedding_technique.py
@@ -13,7 +13,7 @@ if TYPE_CHECKING:
     from clayrs.content_analyzer.embeddings.embedding_loader.embedding_loader import \
         WordEmbeddingLoader, SentenceEmbeddingLoader, DocumentEmbeddingLoader
     from clayrs.content_analyzer.embeddings.embedding_loader.embedding_loader import EmbeddingSource
-    from clayrs.content_analyzer.information_processor.information_processor import InformationProcessor
+    from clayrs.content_analyzer.information_processor.information_processor_abstract import InformationProcessor
     from clayrs.content_analyzer.raw_information_source import RawInformationSource
     from clayrs.content_analyzer.information_processor.visualpostprocessor import VisualPostProcessor
