+++ clayrs/content_analyzer/field_content_production_techniques/synset_document_frequency.py
@@ -4,7 +4,7 @@ from sklearn.feature_extraction.text import CountVectorizer
 from typing import List, TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from clayrs.content_analyzer.information_processor.information_processor import InformationProcessor
+    from clayrs.content_analyzer.information_processor.information_processor_abstract import InformationProcessor
     from clayrs.content_analyzer.raw_information_source import RawInformationSource
 
 from clayrs.content_analyzer.field_content_production_techniques.field_content_production_technique import \
