+++ clayrs/content_analyzer/field_content_production_techniques/tf_idf.py
@@ -5,7 +5,7 @@ from sklearn.feature_extraction.text import TfidfVectorizer
 from typing import List, Union, Mapping, Iterable, Callable, TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from clayrs.content_analyzer.information_processor.information_processor import InformationProcessor
+    from clayrs.content_analyzer.information_processor.information_processor_abstract import InformationProcessor
     from clayrs.content_analyzer.raw_information_source import RawInformationSource
 
 from clayrs.content_analyzer.field_content_production_techniques.field_content_production_technique \
