+++ clayrs/content_analyzer/information_processor/spacy_processor.py
@@ -8,7 +8,7 @@ with warnings.catch_warnings():
     import spacy
     from spacy.tokens import Token
 
-from clayrs.content_analyzer.information_processor.information_processor import NLP
+from clayrs.content_analyzer.information_processor.information_processor_abstract import NLP
 from clayrs.content_analyzer.utils.check_tokenization import check_not_tokenized
